Revert "add xcl2"
This reverts commit d318ab5af4c094923d7f371362b7015a385d7f78.
